Code sync.

+++ src/main/java/com/example/taskmanager/exception/UserNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.taskmanager.exception;
 
 public class UserNotFoundException extends RuntimeException{
 
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(Integer id) {
         super("User with id " + id + " not found");
     }
 }
